+++ src/pujaSah/assignments11_20/Last3NumberDivisibilityBy3And7.java
@@ -1,4 +1,4 @@
-package pujaSah;
+package pujaSah.assignments11_20;
 
 public class Last3NumberDivisibilityBy3And7 {
