+++ src/pujaSah/assignments11_20/MethodsOnPrimeNumber.java
@@ -1,4 +1,4 @@
-package pujaSah;
+package pujaSah.assignments11_20;
 
 public class MethodsOnPrimeNumber {
 	int count = 0;
