+++ src/pujaSah/assignments11_20/NumberDivisibilityBy3And7.java
@@ -1,4 +1,4 @@
-package pujaSah;
+package pujaSah.assignments11_20;
 
 public class NumberDivisibilityBy3And7 {
