+++ src/pujaSah/basicJavaConcepts/Calculator.java
@@ -1,4 +1,4 @@
-package pujaSah;
+package pujaSah.basicJavaConcepts;
 
 class Calculator{
 	int number1=10;
