+++ src/pujaSah/basicJavaConcepts/CountNumbers.java
@@ -1,4 +1,4 @@
-package pujaSah;
+package pujaSah.basicJavaConcepts;
 
 class CountNumbers{
