+++ src/pujaSah/basicJavaConcepts/Example1.java
@@ -1,4 +1,4 @@
-package pujaSah;
+package pujaSah.basicJavaConcepts;
 
 class Example1{
