+++ src/pujaSah/basicJavaConcepts/Example2.java
@@ -1,4 +1,4 @@
-package pujaSah;
+package pujaSah.basicJavaConcepts;
 
 class Example2{
