+++ src/pujaSah/sssignments1_10/InitialAssignmentBankDetail.java
@@ -1,4 +1,4 @@
-package pujaSah;
+package pujaSah.sssignments1_10;
 
 public class InitialAssignmentBankDetail {
 	int balance = 10000;
