+++ src/pujaSah/sssignments1_10/InitialAssignmentCalculator3.java
@@ -1,4 +1,4 @@
-package pujaSah;
+package pujaSah.sssignments1_10;
 
 public class InitialAssignmentCalculator3 {
