+++ src/pujaSah/sssignments1_10/InitialAssignmentNumberAddition.java
@@ -1,4 +1,4 @@
-package pujaSah;
+package pujaSah.sssignments1_10;
 
 public class InitialAssignmentNumberAddition {
 	void processData(int firstIndex, int lastIndex) {
