+++ src/pujaSah/sssignments1_10/InitialAssignmentNumberSystem.java
@@ -1,4 +1,4 @@
-package pujaSah;
+package pujaSah.sssignments1_10;
 
 public class InitialAssignmentNumberSystem {
